login

+++ ms-user/src/main/java/com/amx/jax/userservice/repository/CustomerRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.amx.jax.dbmodel.Customer;
 
 @Transactional
-public interface CustomerRepository extends CrudRepository<Customer, Long> {
+public interface CustomerRepository extends CrudRepository<Customer, BigDecimal> {
 
 	@Query("select c from Customer c where countryId=?1 and  identityInt =?2 and isActive='Y'")
 	public List<Customer> getCustomerbyuser(BigDecimal countryId, String userId);
